+++ .size-limit.cjs
@@ -5,7 +5,35 @@ module.exports = [
     import: '{ DotLottie }',
   },
   {
-    name: '@lottiefiles/dotlottie-web WASM',
+    name: '@lottiefiles/dotlottie-web/webgpu',
+    path: 'packages/web/dist/webgpu/index.js',
+    import: '{ DotLottie }',
+  },
+  {
+    name: '@lottiefiles/dotlottie-web/webgl',
+    path: 'packages/web/dist/webgl/index.js',
+    import: '{ DotLottie }',
+  },
+  {
+    name: '@lottiefiles/dotlottie-web WASM (WebGL)',
+    path: 'packages/web/dist/webgl/*.wasm',
+    modifyWebpackConfig: (config) => {
+      config.experiments = {
+        asyncWebAssembly: true,
+      };
+    },
+  },
+  {
+    name: '@lottiefiles/dotlottie-web WASM (WebGPU)',
+    path: 'packages/web/dist/webgpu/*.wasm',
+    modifyWebpackConfig: (config) => {
+      config.experiments = {
+        asyncWebAssembly: true,
+      };
+    },
+  },
+  {
+    name: '@lottiefiles/dotlottie-web WASM (Software)',
     path: 'packages/web/dist/*.wasm',
     modifyWebpackConfig: (config) => {
       config.experiments = {