+++ packages/web/src/core/index.ts
@@ -1,2 +1,0 @@
-export type * from './dotlottie-player.types';
-export * from './dotlottie-wasm-loader';