+++ packages/web/tests/setup.ts
@@ -1,4 +1,0 @@
-import { DotLottie } from '../src';
-import wasmUrl from '../src/core/dotlottie-player.wasm?url';
-
-DotLottie.setWasmUrl(wasmUrl);